+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.8.0
+
+* `PersistenceImage` changes:
+  - `slope_end` is now relative to top coordinate in image.
+  - default value of `sigma` changed to 2× pixel size
+  - improved performance.
+
 # v0.7.3
 
 * Added Tables.jl interface